+++ src/features/flags/types.ts
@@ -204,6 +204,7 @@ export type FeatureFlag =
   | 'switchCircuits'
   | 'settingOutDims'
   | 'lampSpecChecks'
+  | 'variationRegister'
   | 'tileLayoutSheet'
   | 'carpentrySheets'
   | 'rcpSheet'